sw_services: xilplmi: Fixes for MISRA C violation Rule 10.4
MISRA C Rule 10.4 States: Both operands of an operator in which the usual
arithmetic conversionsare performed shall have the same essential type category

This patch fixes MISRA C Rule 10.4 violation by using the same type variable
for both operands where possible


Acked-for-series: Vikram Sreenivasa Batchali <b.vikram@xilinx.com>

+++ lib/sw_services/xilplmi/src/xplmi_cdo.h
@@ -75,8 +75,8 @@ typedef struct {
 	u32 TempCmdBuf[8U];	/**< Temporary buffer to store commands
 				 between iterations */
 	XPlmi_Cmd Cmd;		/**< Pointer to the cmd */
-	u32 CmdEndDetected;	/**< Flag to detect end of commands */
-	u32 Cdo1stChunk;	/**< This is used for first time to validate
+	u8 CmdEndDetected;	/**< Flag to detect end of commands */
+	u8 Cdo1stChunk;		/**< This is used for first time to validate
 				CDO header*/
 	u32 ImgId;		/**< Info about which Image this belongs to */
 	u32 PrtnId;		/**< Info about which partition this belongs to*/
